# api/gen/models/setting.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
)

// Setting User stock setting
// swagger:model Setting

type Setting struct {

	// Key
	Key string `json:"key,omitempty"`

	// Value
	Value string `json:"value,omitempty"`
}

/* polymorph Setting key false */

/* polymorph Setting value false */

// Validate validates this setting
func (m *Setting) Validate(formats strfmt.Registry) error {
	var res []error

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

// MarshalBinary interface implementation
func (m *Setting) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Setting) UnmarshalBinary(b []byte) error {
	var res Setting
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}