# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "Stock Assistant",
    "contact": {
      "name": "mars"
    },
    "version": "v1"
  },
  "basePath": "/api/stock-assistant/v1",
  "consumes": [
    "application/json;charset=utf-8"
  ],
  "produces": [
    "application/json;charset=utf-8"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "responses": {
    "InternalServerError":{
      "description": "Internal server error",
      "schema": {
        "type": "string"
      }
    }
  },
  "paths": {
    "/{userId}/indices": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "summary": "Get user indices",
        "operationId": "UserIndexList",
        "security": [
          {
          }
        ],
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/StockIndex"
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "save",
        "operationId": "UserIndexSave",
        "parameters": [
          {
            "in": "body",
            "name": "Index",
            "description": "Index",
            "schema": {
              "$ref": "#/definitions/StockIndex"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/StockIndex"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/{userId}/indices/rename":{
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        }
      ],
      "post": {
        "summary": "",
        "operationId": "UserIndexRename",
        "parameters": [
          {
            "in": "query",
            "name": "oldName",
            "description": "old name",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "newName",
            "description": "new name",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/StockIndex"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/{userId}/indices/{indexId}": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        },
        {
          "in": "path",
          "name": "indexId",
          "description": "index id",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "summary": "Get user index",
        "operationId": "UserIndexGet",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/StockIndex"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "",
        "operationId": "UserIndexDelete",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/{userId}/stockEvaluates": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "UserStockEvaluateList",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/StockEvaluate"
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "save",
        "operationId": "UserStockEvaluateSave",
        "parameters": [
          {
            "in": "body",
            "name": "stockEvaluate",
            "description": "",
            "schema": {
              "$ref": "#/definitions/StockEvaluate"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/StockEvaluate"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/{userId}/stockEvaluates/{stockId}": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        },
        {
          "in": "path",
          "name": "stockId",
          "description": "stock id",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "UserStockEvaluateGet",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/StockEvaluate"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/{userId}/stockEvaluates/{stockId}/indexEvaluates": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        },
        {
          "in": "path",
          "name": "stockId",
          "description": "stock id",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "UserIndexEvaluateList",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IndexEvaluate"
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "",
        "operationId": "UserIndexEvaluateSave",
        "parameters": [
          {
            "in": "body",
            "name": "indexEvaluate",
            "schema": {
              "$ref": "#/definitions/IndexEvaluate"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/IndexEvaluate"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/{userId}/stockEvaluates/{stockId}/indexEvaluates/{indexName}": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        },
        {
          "in": "path",
          "name": "stockId",
          "description": "stock id",
          "type": "string",
          "required": true
        },
        {
          "in": "path",
          "name": "indexName",
          "description": "index name",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "UserIndexEvaluateGet",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/IndexEvaluate"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/{userId}/settings": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "summary": "list",
        "operationId": "UserSettingList",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Setting"
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "save",
        "operationId": "UserSettingSave",
        "parameters": [
          {
            "in": "body",
            "name": "Setting",
            "description": "setting",
            "schema": {
              "$ref": "#/definitions/Setting"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/Setting"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/{userId}/settings/{configKey}": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "description": "User id",
          "type": "string",
          "required": true
        },
        {
          "in": "path",
          "name": "configKey",
          "description": "config key",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "UserSettingGet",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/Setting"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "",
        "operationId": "UserSettingDelete",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    }
  },
  "definitions": {
    "Setting": {
      "id": "Setting",
      "description": "User stock setting",
      "type": "object",
      "properties": {
        "key": {
          "description": "Key",
          "type": "string"
        },
        "value": {
          "description": "Value",
          "type": "string"
        }
      }
    },
    "StockIndex": {
      "id": "Index",
      "description": "User stock index",
      "type": "object",
      "properties": {
        "name": {
          "description": "name",
          "type": "string"
        },
        "desc": {
          "description": "desc",
          "type": "string"
        },
        "evalWeight": {
          "description": "Eval weight",
          "type": "string",
          "format": "int32"
        },
        "aiWeight": {
          "description": "ai weight",
          "type": "string",
          "format": "int32"
        },
        "niWeight": {
          "description": "ni weight",
          "type": "string",
          "format": "int32"
        }
      }
    },
    "StockEvaluate": {
      "id": "StockEvaluate",
      "description": "stock evaluate",
      "type": "object",
      "properties": {
        "stockId": {
          "description": "stock id",
          "type": "string"
        },
        "totalScore": {
          "description": "score",
          "type": "string",
          "format": "double"
        },
        "evalRemark": {
          "description": "remark",
          "type": "string"
        }
      }
    },
    "IndexEvaluate": {
      "id": "IndexEvaluate",
      "description": "index evaluate",
      "type": "object",
      "properties": {
        "indexName": {
          "description": "index name",
          "type": "string"
        },
        "evalStars": {
          "description": "eval stars",
          "type": "string",
          "format": "int32"
        },
        "evalRemark": {
          "description": "eval remark",
          "type": "string"
        },
        "updateTime": {
          "description": "update time",
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}